move declaration of L1RpcServer into ch_frb_rpc.hpp; remove old definition
